Dialogs show_xlog now remember the offset position in log file as off_t instead of int. Otherwise with large log data we have an overflow.

+++ src/UI/Motif/show_dlog/show_dlog.h
@@ -1,6 +1,6 @@
 /*
  *  show_dlog.h - Part of AFD, an automatic file distribution program.
- *  Copyright (c) 1998 - 2023 Holger Kiehl <Holger.Kiehl@dwd.de>
+ *  Copyright (c) 1998 - 2026 Holger Kiehl <Holger.Kiehl@dwd.de>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -84,11 +84,11 @@
 /* Structure that holds offset (to job ID) to each item in list. */
 struct item_list
        {
-          FILE *fp;
-          int  no_of_items;
-          int  *line_offset;   /* Array that contains the offset to the */
+          FILE  *fp;
+          int   no_of_items;
+          off_t *line_offset;  /* Array that contains the offset to the */
                                /* file name of that item.               */
-          int  *offset;        /* Array that contains the offset to the */
+          off_t *offset;       /* Array that contains the offset to the */
                                /* ID of that item.                      */
        };
